#5094 Move to repo checking
#5094 Extend import process
