+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/impexp/DBFunctions.java
@@ -3,15 +3,10 @@
  */
 package org.adempiere.impexp;
 
-import java.util.ArrayList;
 import java.util.List;
-
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
 
 import com.google.common.collect.ImmutableList;
 
-import de.metas.logging.LogManager;
 import lombok.Builder;
 import lombok.Builder.Default;
 import lombok.NonNull;
@@ -47,38 +42,10 @@ import lombok.Value;
 @Value
 public class DBFunctions
 {
-	private static final transient Logger log = LogManager.getLogger(DBFunctions.class);
-	private final static String IMPORT_AFTER_ROW = "IMPORT_AFTER_ROW";
-
 	@NonNull
 	private final String tableName;
 	
 	@NonNull
 	@Default
-	private final List<DBFunction> importFunctions = ImmutableList.of();
-
-	public List<DBFunction> getEliglibleImportFunctions()
-	{
-		final List<DBFunction> availableAfterRowFunctions = new ArrayList<>();
-		for (final DBFunction function : importFunctions)
-		{
-			if (isEligibleFunction(function))
-			{
-				availableAfterRowFunctions.add(function);
-			}
-			else
-			{
-				log.warn("Function {} from schema {} is not eliglible for importing process!", function.getName(), function.getSchema());
-			}
-		}
-		
-		return ImmutableList.copyOf(availableAfterRowFunctions);
-	}
-
-	private boolean isEligibleFunction(@NonNull final DBFunction function)
-	{
-		final String routine_name = function.getName();
-		return StringUtils.containsIgnoreCase(routine_name, IMPORT_AFTER_ROW);
-	}
-
+	private final List<DBFunction> availableAfterRowFunctions = ImmutableList.of();
 }
